requirements/base.txt
# Core
numpy
pydantic
python-dotenv
rich
loguru
keyboard

# Async / orchestration
anyio
asyncio-mqtt